unkempt guide: fix up / clarify uses of "git checkout -b"
The "-b" option of "git checkout" is not a general behavioral modifier
that makes the command interpret the first operand as the new branch to
create. Instead, the "-b" option takes the new branch directly as its
argument.

Some uses of "-b" are incorrect in the arcticle, fix them; some others
happen to be correct, but their layout is misleading, reformat those.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>